backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt<5]
celery
redis
python-multipart
email-validator